Generalized erase to be part of the crtp
The erase methods are now in the crtp base class, and handles the
erasure from the trie. Inheriting classes implement an invalidate_node
callback method to maintain their additional node information. For
example: the basic_trie_set implements it by removing the node from its
set of members.

Also made the crop base class a friend of the implementing classes, so
that methods can be hidden from the user, but used by the crtp base
(currently is_valid_node (formerly known as validate) and
invalidate_node).

The crtp function force_node changed name to insert_node, which is more
consistent with the container semantics.