Added trie_map
Added a trie_map class that associates inserted keys with mapped values.

This made it necessary to add functionality to have the implementing
trie interpret a node so that meaningful values can be dereferenced
from the iterators (node_to_ref and note_to_ptr). The current
implementation is not optimal, since each dereferencing essentially
does an additional validation of whether the node is associated with a
value or not. It already had to do such a check when validating during
traversal, which is wasted.

Also: the iterators now have to differentiate their pointers to
containing trie as either mutable or immutable, otherwise there is no
way to have mapped values be altered from a non-const iterator (and
also it just wont compile - we need const correctness).

I’m very unhappy with the node_to_ptr implementation for trie_map, but
at least it works.

Split the test program so that basic functionality is tested in
trie_test, whereas the set and map implementations get their own test
files. This helps me stay sane.