Trie set iterators now dereference to a key range
The iterators of trie set now dereference to a range of elements, which
serves as approximately a key. The problem now is that a new
trie_path_t is created every time the iterator is dereferenced, so a
begin-end for loop with two dereferences of an iterator will fail
(since the begin and end instantiate two different paths, producing an
undefined range).

Factored out the trie_path_t into its own header.